Fixed import issues and refactored util functions
