+++ vscode/src/lsp/listeners/requests/handlers.ts
@@ -14,14 +14,15 @@
   limitations under the License.
 */
 import { QuickPickItem, Uri, window, workspace, WorkspaceConfiguration } from "vscode";
-import { globalVars, LOGGER } from "../../extension";
-import { notificationOrRequestListenerType } from "../types";
-import { ExecInHtmlPageRequest, HtmlPageRequest, InputBoxRequest, InputBoxStep, MutliStepInputRequest, QuickPickRequest, QuickPickStep, SaveDocumentRequestParams, SaveDocumentsRequest, TextEditorDecorationCreateRequest, UpdateConfigurationRequest } from "../protocol";
-import { InputStep, MultiStepInput } from "../../utils";
-import { runConfigurationUpdateAll } from "../../runConfiguration";
-import { isError, isString } from "../../typesUtil";
-import { LogLevel } from "../../logger";
-import { execInHtmlPage, showHtmlPage } from "../../webviews/nbWebviewHandler";
+import { globalVars, LOGGER } from "../../../extension";
+import { notificationOrRequestListenerType } from "../../types";
+import { ExecInHtmlPageRequest, HtmlPageRequest, InputBoxRequest, InputBoxStep, MutliStepInputRequest, QuickPickRequest, QuickPickStep, SaveDocumentRequestParams, SaveDocumentsRequest, TextEditorDecorationCreateRequest, UpdateConfigurationRequest } from "../../protocol";
+import { InputStep, MultiStepInput } from "../../../utils";
+import { runConfigurationUpdateAll } from "../../../runConfiguration";
+import { isError } from "../../../utils";
+import { isString } from "../../../utils";
+import { LogLevel } from "../../../logger";
+import { execInHtmlPage, showHtmlPage } from "../../../webviews/nbWebviewHandler";
 
 const textEditorDecorationCreateRequestHandler = (param: any) => {
     let decorationType = window.createTextEditorDecorationType(param);
